Implement UpdateTimestamps for Linux (untested)
