+++ src/shared/platform.h
@@ -12,6 +12,12 @@
 #define PRFILESYSTEM_PATH "s"
 #endif
 
+namespace cd
+{
+	struct ISO_DATESTAMP;
+}
+
 FILE* OpenFile(const std::filesystem::path& path, const char* mode);
 std::optional<struct _stat64> Stat(const std::filesystem::path& path);
 int64_t GetSize(const std::filesystem::path& path);
+void UpdateTimestamps(const std::filesystem::path& path, const cd::ISO_DATESTAMP& entryDate);
